+++ packages/kofono/src/builder/LeafBuilder.ts
@@ -1,7 +1,6 @@
-import type { Form } from "../form/Form";
+import { normalizeEnumDef } from "../property/enum";
 import { Property } from "../property/Property";
 import type { SchemaProperty } from "../schema/Schema";
-import type { Builder } from "./Builder";
 import type { PropertyBuilder } from "./types";
 
 export class LeafBuilder<TSchemaType extends SchemaProperty>
@@ -9,15 +8,18 @@ export class LeafBuilder<TSchemaType extends SchemaProperty>
 {
     public constructor(
         protected readonly uid: string,
-        protected readonly builder: Builder,
         protected def: TSchemaType,
-    ) {}
-
-    async build(): Promise<Form> {
-        return await this.builder.build();
+    ) {
+        this.normalizeEnum();
     }
 
-    buildProperty(): Property<TSchemaType> {
+    build(): Property<TSchemaType> {
         return new Property<TSchemaType>(this.uid, this.def);
+    }
+
+    private normalizeEnum() {
+        if (this.def.enum) {
+            this.def.enum = normalizeEnumDef(this.def.enum);
+        }
     }
 }