+++ packages/kofono/src/builder/NodeBuilder.ts
@@ -1,8 +1,5 @@
-import type { Form } from "../form/Form";
 import { Property } from "../property/Property";
-import type { SchemaObjectProperty, SchemaProperty } from "../schema/Schema";
-import { joinSelectors } from "../selector/helpers";
-import type { Builder } from "./Builder";
+import type { SchemaProperty } from "../schema/Schema";
 import type { PropertyBuilder } from "./types";
 
 export class NodeBuilder<TSchemaType extends SchemaProperty>
@@ -10,26 +7,10 @@ export class NodeBuilder<TSchemaType extends SchemaProperty>
 {
     public constructor(
         protected readonly uid: string,
-        protected readonly builder: Builder,
         protected def: TSchemaType,
     ) {}
 
-    object(
-        uid: string,
-        def: Omit<SchemaObjectProperty, "type">,
-    ): NodeBuilder<SchemaObjectProperty> | null {
-        const objectDef = {
-            type: "object",
-            ...def,
-        };
-        return this.builder.object(joinSelectors(this.uid, uid), objectDef);
-    }
-
-    async build(): Promise<Form> {
-        return await this.builder.build();
-    }
-
-    buildProperty(): Property<TSchemaType> {
+    build(): Property<TSchemaType> {
         return new Property<TSchemaType>(this.uid, this.def);
     }
 }