+++ packages/kofono/CHANGELOG.md
@@ -3,6 +3,8 @@
 - add cache to speed up form stats compilation
 - remove YAML schema support
 - replace date-fns dependency by local functions
+- refactor property builder
+- add schema normalization process at build stage
 
 # 0.8.0 - 2026-03-29